refactor: Restructure application and improve documentation
- Restructure app into modular components for better organization
- Update project documentation with new structure and module descriptions
- Improve installation and Google Cloud setup instructions
- Implement environment variable loading and consistent PROJECT_ID usage
