+++ requirements.txt
@@ -1,0 +1,78 @@
+altair==5.5.0
+annotated-types==0.7.0
+anyio==4.9.0
+attrs==25.3.0
+blinker==1.9.0
+branca==0.8.1
+cachetools==5.5.2
+certifi==2025.1.31
+charset-normalizer==3.4.1
+click==8.1.8
+db-dtypes==1.4.2
+docstring_parser==0.16
+folium==0.19.5
+geopandas==1.0.1
+gitdb==4.0.12
+GitPython==3.1.44
+google-api-core==2.24.2
+google-auth==2.38.0
+google-cloud-aiplatform==1.87.0
+google-cloud-bigquery==3.31.0
+google-cloud-bigquery-storage==2.30.0
+google-cloud-core==2.4.3
+google-cloud-resource-manager==1.14.2
+google-cloud-storage==2.19.0
+google-crc32c==1.7.1
+google-genai==1.9.0
+google-resumable-media==2.7.2
+googleapis-common-protos==1.69.2
+grpc-google-iam-v1==0.14.2
+grpcio==1.71.0
+grpcio-status==1.71.0
+h11==0.14.0
+httpcore==1.0.7
+httpx==0.28.1
+idna==3.10
+Jinja2==3.1.6
+jsonschema==4.23.0
+jsonschema-specifications==2024.10.1
+MarkupSafe==3.0.2
+narwhals==1.33.0
+numpy==2.2.4
+packaging==24.2
+pandas==2.2.3
+pillow==11.1.0
+proto-plus==1.26.1
+protobuf==5.29.4
+pyarrow==19.0.1
+pyasn1==0.6.1
+pyasn1_modules==0.4.2
+pydantic==2.11.2
+pydantic_core==2.33.1
+pydeck==0.9.1
+pyogrio==0.10.0
+pyproj==3.7.1
+python-dateutil==2.9.0.post0
+python-dotenv==1.1.0
+pytz==2025.2
+referencing==0.36.2
+requests==2.32.3
+rpds-py==0.24.0
+rsa==4.9
+shapely==2.1.0
+six==1.17.0
+smmap==5.0.2
+sniffio==1.3.1
+streamlit==1.44.1
+streamlit-chat==0.1.1
+streamlit_folium==0.24.0
+tenacity==9.1.2
+toml==0.10.2
+tornado==6.4.2
+typing-inspection==0.4.0
+typing_extensions==4.13.0
+tzdata==2025.2
+urllib3==2.3.0
+watchdog==6.0.0
+websockets==15.0.1
+xyzservices==2025.1.0
